+++ cogs/Verification.py
@@ -25,8 +25,8 @@ from utils import KoalaDBManager
 
 # Constants
 load_dotenv()
-GMAIL_EMAIL = os.environ['GMAIL_EMAIL']
-GMAIL_PASSWORD = os.environ['GMAIL_PASSWORD']
+GMAIL_EMAIL = os.environ.get('GMAIL_EMAIL')
+GMAIL_PASSWORD = os.environ.get('GMAIL_PASSWORD')
 # Variables
 
 
@@ -51,7 +51,7 @@ class Verification(commands.Cog, name="Verify"):
     def __init__(self, bot, db_manager=None):
         self.bot = bot
         if not db_manager:
-            self.DBManager = KoalaDBManager.KoalaDBManager(KoalaBot.DATABASE_PATH, KoalaBot.DB_KEY)
+            self.DBManager = KoalaBot.database_manager
             self.set_up_tables()
             self.DBManager.insert_extension("Verify", 0, True, True)
         else:
@@ -435,11 +435,15 @@ This email is stored so you don't need to verify it multiple times across server
                 print(f"user with id {user_id} not found in {guild}")
 
 
-
 def setup(bot: KoalaBot) -> None:
     """
     Load this cog to the KoalaBot.
     :param bot: the bot client for KoalaBot
     """
-    bot.add_cog(Verification(bot))
-    print("Verification is ready.")
+    if GMAIL_EMAIL is None or GMAIL_PASSWORD is None:
+        print("Verification not started. API keys not found in environment.")
+        KoalaBot.database_manager.insert_extension("Verify", 0, False, False)
+    else:
+        bot.add_cog(Verification(bot))
+        print("Verification is ready.")
+