+++ CHANGELOG.md
@@ -3,11 +3,13 @@ All notable changes to KoalaBot will be documented in this file.
 A lot of these commands will only be available to administrators
 
 ## [Unreleased]
-###Announce
+### Announce
 - Announce is now enabled successfully
+### Other
+- KoalaBot can now be started without Twitch or Gmail keys (disabling TwitchAlert and Verify)
 
 ## [0.4.0] - 06-04-2021
-###Announce
+### Announce
 ##### Added
 - `announce create` Start the creation of an announcement
 - `announce changeTitle` Changes the title of a pending announcement